faq-chatbot/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
openai==0.28.0
python-dotenv==1.0.1